Change MailboxLock to an interface that is produced by MailboxLockFactory (another interface).
- Use AutoClosable so that MailboxLock instances can be cleaned up via try-with-resources
- This is a step towards allowing the locking mechanism to be injected into the mailbox
- We want to be able to inject the locking mechanism so that in a highly available configuration we can use a distributed lock provider like Redis
