+++ common/src/java/com/zimbra/common/mailbox/MailboxLock.java
@@ -1,0 +1,15 @@
+package com.zimbra.common.mailbox;
+
+public interface MailboxLock extends AutoCloseable {
+    boolean isWriteLock();
+
+    boolean isWriteLockedByCurrentThread();
+
+    boolean isUnlocked();
+
+    int getHoldCount();
+
+    void lock();
+
+    void close();
+}
