+++ common/src/java/com/zimbra/common/mailbox/MailboxLockFactory.java
@@ -1,0 +1,10 @@
+package com.zimbra.common.mailbox;
+
+public interface MailboxLockFactory extends AutoCloseable {
+    MailboxLock readLock();
+
+    MailboxLock writeLock();
+
+    @Deprecated
+    MailboxLock lock(boolean write);
+}
